fix(trusty-mpm): cap Telegram chat history + skip empty assistant turns (review #1501)
Address trusty-review findings before merge:

- MEDIUM: bound the per-chat coordinator history with a sliding window
  (MAX_CHAT_HISTORY_TURNS = 20 messages = last 10 exchanges). Extract a
  record_chat_turn helper that pushes the turn then drains the front so the
  stored history never grows without bound (flat memory, relevant context).
- LOW: skip storing an empty assistant reply — only the user turn is pushed
  when the coordinator returns an empty reply, avoiding context pollution.
- LOW: tighten coordinator_chat_body visibility from pub(crate) to private;
  the sole call site and the test both live within the http_client module.

Split the telegram inline test module into a sibling tests.rs so mod.rs
stays under the 500-SLOC production cap (the added helper pushed it to 521);
tests.rs is a recognized test-file path under the 1500-SLOC test cap. Add
record_chat_turn_caps_history and record_chat_turn_skips_empty_reply.
